+++ api/docs/docs/contact.md
@@ -1,0 +1,6 @@
+---
+id: contact
+title: Contact Us
+---
+
+## Email Support
