+++ api/docs/docs/getting-started.md
@@ -1,0 +1,4 @@
+---
+id: getting-started
+title: Getting Started
+---
