+++ api/docs/docs/overview.md
@@ -1,0 +1,20 @@
+---
+id: overview
+title: Overview
+sidebar_label: Overview
+slug: /
+---
+
+The Covid Act Now API provides the same data that powers [CovidActNow.org](https://covidactnow.org) but in an easily digestible, machine readable format, intended for consumption by other COVID websites, models, and tools. 
+
+### Update frequency
+
+Data is updated every day, typically around noon US Pacific Time.
+
+### Rate Limits
+
+There are no rate limits, but please be aware of your usage as we're a non-profit and would like to stay available to everyone.
+
+### License
+
+Data is licensed under [Creative Commons 4.0 By Attribution](https://creativecommons.org/licenses/by/4.0/). You are welcome to share, copy, and redistribute it, as well as adapt it for your own works, we just ask that you provide attribution to the source (as we have done with [our data sources](https://github.com/covid-projections/covid-data-public#date-sources-for-current--future-use)).
