+++ api/docs/docusaurus.config.js
@@ -15,12 +15,14 @@ module.exports = {
         src: "img/can_logo.png",
       },
       items: [
+       
         {
-          to: "docs/",
-          activeBasePath: "docs",
-          label: "Docs",
+          to: "api",
+          activeBasePath: "api",
+          label: "API Reference",
           position: "left",
         },
+
         {
           href: "https://github.com/covid-projections/covid-data-model",
           label: "GitHub",
@@ -80,16 +82,12 @@ module.exports = {
       {
         docs: {
           sidebarPath: require.resolve("./sidebars.js"),
+          routeBasePath: "/",
+
           // Please change this to your repo.
-          editUrl:
-            "https://github.com/facebook/docusaurus/edit/master/website/",
+          
         },
-        blog: {
-          showReadingTime: true,
-          // Please change this to your repo.
-          editUrl:
-            "https://github.com/facebook/docusaurus/edit/master/website/blog/",
-        },
+
         theme: {
           customCss: require.resolve("./src/css/custom.css"),
         },
