+++ api/docs/package.json
@@ -17,6 +17,7 @@
     "clsx": "^1.1.1",
     "core-js": "^3.6.5",
     "mobx": "^4.2.0",
+    "openapi-markdown": "^1.2.3",
     "react": "^16.8.4",
     "react-dom": "^16.13.1",
     "redoc": "^2.0.0-rc.40",
